Refactor date and time formatting
Pull code out of events-related files into a separate file in the `lib/` directory.

Require this file inside of events/events.js and events/ui.js.
Use functions that are exported from this file to properly format date and time.
