+++ assets/scripts/events/events.js
@@ -1,23 +1,10 @@
 'use strict';
 
 const getFormFields = require('../../../lib/get-form-fields');
+const formatDateTime = require('../../../lib/format-date-time');
 
 const api = require('./api');
 const ui = require('./ui');
-//const app = require('../app');
-
-const formatTime = (data) => {
-  // this is string interpolation using "template literals"
-  // same as:
-  // data.event.data + 'T' + data.event.startTime + ':00'
-  if (data.event.startTime) {
-    data.event.startTime = `${data.event.date}T${data.event.startTime}:00`;
-  }
-  if (data.event.endTime) {
-    data.event.endTime = `${data.event.date}T${data.event.endTime}:00`;
-  }
-  return data;
-};
 
 // for getting all events (button)
 
@@ -51,7 +38,7 @@ const createEvent = function (event) {
   event.preventDefault();
   let data = getFormFields(event.target);
   console.log(data);
-  data = formatTime(data);
+  data = formatDateTime.formatTimeForMongo(data);
   console.log('improved', data);
   api.createNewEvent(data)
     .done(ui.createEventsSuccess)
@@ -72,7 +59,7 @@ const getEditForm = function (event) {
 const updateEvent = function (event) {
   event.preventDefault();
   let data = getFormFields(event.target);
-  data = formatTime(data);
+  data = formatDateTime.formatTimeForMongo(data);
   api.updateEvent(data)
     .done(ui.singleEventSuccess)
     .fail(ui.failure);
